refactor : 카드 관련 컴포넌트 분리
